Implement outer/inner drop region indicator
There are two drop regions per direction.

Inner drop region behaves like move to mark.

The outer region is close to the edge (currently 30px from the edge).
This will place the container as a sibling in the given direction
within the parent container.
If the move direction goes against the orientation of the parent container, the con moves out of the row.

Co-authored-by: Orestis Floros <orestisf1993@gmail.com>
(inner region like move to mark)